web/requirements.txt
appdirs==1.4.4
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
cssselect==1.1.0
Django==4.0.4
django-better-admin-arrayfield==1.4.2
django-bulk-update-or-create==0.3.0
djangorestframework==3.13.1
fake-useragent==0.1.11
idna==3.3
importlib-metadata==4.11.3
lxml==4.8.0
parse==1.19.0
psycopg2-binary==2.9.3
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
pytz==2022.1
requests==2.27.1
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.2
tqdm==4.64.0
urllib3==1.26.9
w3lib==1.22.0
websockets==10.2
zipp==3.8.0
